split player into components
