+++ client/src/app/player/player.component.html
@@ -7,73 +7,9 @@
 </div>
 <div id="container">
   <div class="component">
-    <div id="imageContainer">
-      <img [src]="queueInfo?.currentTrack?.thumbnailUrl">
-    </div>
-    <div id="currentTrackInfo" class="shadow-bottom sticky-top">
-      <h3>{{queueInfo?.currentTrack?.title}}</h3>
-      <small>{{queueInfo?.currentTrack?.artist}}</small>
-      <small>{{queueInfo?.currentTrack?.position | minuteSeconds}}
-        /{{queueInfo?.currentTrack?.duration | minuteSeconds}}</small>
-      <div class="buttons">
-        <button mat-icon-button class="dummy">
-        </button>
-        <button mat-icon-button class="dummy">
-        </button>
-        <button mat-icon-button class="dummy">
-        </button>
-        <div class="filler"></div>
-        <button mat-icon-button (click)="skipBack()">
-          <mat-icon>skip_previous</mat-icon>
-        </button>
-        <button mat-icon-button (click)="togglePause()">
-          <mat-icon>{{queueInfo?.currentTrack?.paused ? 'play_arrow' : 'pause'}}</mat-icon>
-        </button>
-        <button mat-icon-button (click)="skip()">
-          <mat-icon>skip_next</mat-icon>
-        </button>
-        <div class="filler"></div>
-        <button mat-icon-button (click)="toggleRepeat()">
-          <mat-icon [class]="queueInfo?.repeat ? '' : 'disabled'">repeat</mat-icon>
-        </button>
-        <button mat-icon-button (click)="toggleAutoRadio()">
-          <mat-icon [class]="queueInfo?.autoRadio ? '' : 'disabled'">radio-outline</mat-icon>
-        </button>
-        <button mat-icon-button (click)="clear()">
-          <mat-icon>delete_sweep</mat-icon>
-        </button>
-      </div>
-    </div>
-    <div cdkDropList id="queueList" #queueList="cdkDropList" [cdkDropListData]="queueInfo?.tracks"
-         (cdkDropListDropped)="drop($event)">
-      <app-track-info *ngFor="let track of queueInfo?.tracks" [track]="track" [events]="['REMOVE']"
-                      (event)="trackInfoEvent($event)"></app-track-info>
-    </div>
+    <app-queue [searchResult]="searchResult"></app-queue>
   </div>
   <div class="component">
-    <div id="url-bar-container" class="shadow-bottom sticky-top">
-      <button mat-icon-button *ngIf="isBrowsing" (click)="back()">
-        <mat-icon>arrow_back_ios</mat-icon>
-      </button>
-      <mat-form-field id="urlInput">
-        <mat-label>Search / Url</mat-label>
-        <input matInput #searchElement [(ngModel)]="url" (keyup.enter)="enter()" (keyup.control.enter)="now()"
-               (keyup.control.shift.enter)="next()" (keyup.shift.enter)="queue()"
-               (click)="selectInput($event)">
-      </mat-form-field>
-      <button mat-stroked-button (click)="search()">Search</button>
-      <button mat-stroked-button (click)="now()">Now</button>
-      <button mat-stroked-button (click)="next()">Next</button>
-      <button mat-stroked-button (click)="queue()">Queue</button>
-    </div>
-    <div *ngIf="loading" class="spinner-container">
-      <mat-spinner></mat-spinner>
-    </div>
-    <div *ngIf="!loading" id="searchList" cdkDropList [cdkDropListConnectedTo]="queueList"
-         [cdkDropListSortingDisabled]="true" [cdkDropListData]="searchResult">
-      <app-track-info *ngFor="let track of searchResult" [track]="track"
-                      [events]="['NOW', 'NEXT', 'QUEUE', 'BROWSE', 'RADIO']"
-                      (event)="trackInfoEvent($event)"></app-track-info>
-    </div>
+    <app-search (searchResultChange)="searchResult = $event"></app-search>
   </div>
 </div>
