+++ client/src/app/player/player.component.scss
@@ -37,89 +37,5 @@
     max-height: 100%;
     overflow-y: auto;
   }
-
-  #imageContainer {
-    padding: 0;
-    margin: 0.3rem;
-    max-width: 99%;
-    text-align: center;
-
-    img {
-      max-height: 100%;
-      max-width: 100%;
-    }
-  }
-
-  #currentTrackInfo {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin: 0;
-    padding: 0.3rem;
-    max-width: 99%;
-
-    > * {
-      flex: 1 1 auto;
-      text-align: center;
-    }
-
-    h3 {
-      margin: 0;
-    }
-
-    .buttons {
-      display: flex;
-      flex-wrap: nowrap;
-      justify-content: center;
-
-      > * {
-        flex: 0 0 auto;
-      }
-
-      .filler {
-        flex-grow: 1;
-      }
-
-      .dummy {
-        visibility: hidden;
-      }
-
-      .disabled {
-        color: darkgray;
-      }
-    }
-  }
-
-  #queueList {
-    height: 100%;
-  }
-
-  .spinner-container {
-    display: flex;
-    justify-content: center;
-    margin-top: 2rem;
-  }
-
-  #url-bar-container {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    > * {
-      margin: 0.2rem;
-    }
-
-    mat-form-field {
-      height: 2.7rem;
-    }
-
-    #urlInput {
-      flex: 1 1 auto;
-    }
-
-    button {
-      flex: 0 0 auto;
-      margin-bottom: auto;
-    }
-  }
 }
